Changed database structure
Changed the DB structure so it has a hashed password field and a token field.

I also changed it so the product image field will support file name paths. instead of a blob of data (let's not store the images in the DB).

The users now login with hashes, so the default user data was updated so the passwords are the hash of "test".