Created one Image in Database
Since we have to store an image in the database and load it in the site, I changed the code to import one image into the database on the load script.

NOTE: You have to create a copy of the Cumulonimbus.jpg image in the /var/lib/mysql-files/ directory in order for this command to work. @evgo-space  @ltarr